// src/types.ts
export type UserRole = 'user' | 'support' | 'admin';

export interface Profile {
  id: string;
  email: string;
  name: string;
  username: string; // unique, used for path: sugora.com/u/username
  avatar_url?: string;
  role: UserRole;
  created_at: string;
  is_verified: boolean;
}

export interface Wallet {
  id: string;
  user_id: string;
  balance: number;
  promo_balance: number;
  withdrawn: number;
  updated_at: string;
}

export type TransactionType = 'deposit' | 'withdrawal' | 'affiliate_earning' | 'referral_earning' | 'product_sale' | 'admin_reward';
export type TransactionStatus = 'pending' | 'completed' | 'failed';

export interface WalletTransaction {
  id: string;
  wallet_id: string;
  amount: number;
  type: TransactionType;
  status: TransactionStatus;
  description: string;
  created_at: string;
}